pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repo-cloner"
version = "0.1.0"
description = "Universal Repository Cloner & Synchronization Tool"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Contributors", email = "noreply@github.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "GitPython>=3.1.40",
    "PyGithub>=2.1.1",
    "python-gitlab>=4.4.0",
    "boto3>=1.34.0",
    "azure-storage-blob>=12.19.0",
    "google-cloud-storage>=2.14.0",
    "oci>=2.119.0",
    "PyYAML>=6.0.1",
    "pydantic>=2.5.0",
    "click>=8.1.7",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "packaging>=23.2",
    "toml>=0.10.2",
    "xmltodict>=0.13.0",
    "semver>=3.0.2",
    "tqdm>=4.66.1",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.21.1",
    "moto[s3]>=4.2.9",
    "responses>=0.24.1",
    "requests-mock>=1.11.0",
    "faker>=22.0.0",
    "freezegun>=1.4.0",
    "docker>=7.0.0",
    "testcontainers>=3.7.1",
    "black>=23.12.1",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "isort>=5.13.2",
    "pylint>=3.0.3",
    "pre-commit>=3.6.0",
    "build>=1.0.3",
    "twine>=4.0.2",
]

[project.scripts]
repo-cloner = "repo_cloner.cli:main"

[project.urls]
Homepage = "https://github.com/montge/repo-cloner"
Issues = "https://github.com/montge/repo-cloner/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pylint.messages_control]
max-line-length = 100
disable = ["C0111", "C0103"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v -ra --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["src/repo_cloner"]
omit = ["tests/*", ".venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
